Revert "feat: Setting up Firebase services"
This reverts commit 8723055d4851905375137e329a7fb1f7121fb0c0.

Reverting to working version

+++ updater/src/main/java/kz/market/updater/data/worker/UpdateDownloadWorker.kt
@@ -88,7 +88,7 @@ class UpdateDownloadWorker(
                 )
             )
         } catch (e: IOException) {
-            file.delete()
+            if (file.exists()) file.delete()
 
             if (e.message?.contains("ENOSPC") == true) {
                 return@withContext Result.failure(
